Refactor methods in Main and Controller to not use exceptions, remove use of Observer and add misc. code quality changes.
The methods in Controller and Main, such as start(Stage stage) had 7
different exceptions tied to them, all related to the level loading
system utilizing reflection. Instead of propagating each exception up
to every function in the call stack, all possible exceptions are now
handled in the goToLevel method, as an error window that pauses program
execution and then quits it upon close.

While adding the above, I refactored Levels to be stored as a
Map<String, Class>, and hardcoding the classes at compile time. This
allows for compile-time checks that the levels exist, while still using
String names for levels.

Observer has been deprecated since Java 9, and while there is a modern
equivalent in the form of Listener, I decided to replace it with simply
passing the controller instance to the LevelParent. The reason for
this is that the Observer pattern is most useful for informing many
different parts of the program of an event at once, after which they
can respond asynchronously. In the program, the only way in which this
is used is  informing a single section of the code to load a level,
during which the program should in fact *not* be running asynchronously
in the background. As such, passing the controller instance works as a
much  simpler solution. During this process, I also fixed the loading
of LevelTwo.