Refactor input into InputManager
Creates a new InputManager class that takes a Scene and keeps track of
its KeyUp/KeyDown events. It keeps a map between KeyCodes and
BooleanPropertys that keep track of whether a key is held. It also
provides methods that encapsulate whether an action is being taken,
allowing for multiple keybinds.