Finish refactoring ActiveActor to inherit from Element
Element now has many methods that set its position relative to its
parent, along with moving it a certain amount in a direction. None of
this uses Translate anymore, meaning that everything is done with just
the raw coordinates.

Coordinates are now also from the middle of the image, and starting
positions/bounds are all calculated from the image width. This has
reduced the number of arbitrary constants in UserPlane (for example) by
7.

The shield image is also fixed to the Boss now, albeit its size is
wrong. This removes the need for LevelViewLevelTwo entirely, so it has
been removed.