Refactor project to use a loader and Dependency Injection for resources
Creates an abstract AssetLoader class which contains the a list of
all resources in the project and their paths, and two methods for
loading them (loadImage and loadSound).

Any class in the project that requires resources now gets injected with
an AssetLoader, which it then uses to load resources. This allows assets
to be shared across various Nodes in the JavaFX tree, with customizable
behaviour for the loading.

Comes with two implementations of AssetLoader as an example - an
up-front loader that loads every image on program startup, and a cached
one which loads images and cached them on demand.