Implement the Strategy pattern for FighterPlanes
The strategy pattern is a pattern that allows a custom algorithm to be
selected at runtime, in the form of an instance of a Strategy class.
This allows for loose coupling between the class using the algorithm and
the algorithm itself.

In this case, it means that the logic for the behaviour of the enemies
and player has been moved to separate Strategy classes, making them
highly modular. It would be possible for example, to switch the player's
AI over to that of an enemy by simply changing a single line of code.
A more useful application there, would be being able to dynamically
switch player character behaviour (projectiles fired), or easily make
new enemy types without the need for subclassing (composition over
inheritance).

In the process of implementing the API for the above, I also added a new
Vector class used for the return type of some methods, bundling together
an x and a y value. I would have used this in more places including as
method parameters, if Java supported value types... (C# Structs for
example).