Implement Config classes to streamline object creation
With the refactoring almost done, the one remaining issue is the large
number of constructor arguments present in nearly every class. While
this is a form of code smell, in the form of not following the single
responsibility principle, most of my issues come from the generalization
of actor creation - avoiding creating unnecessary subclasses via passing
parameters such as Heath, Speed, Projectile Y Offset and the Image for
the actor during its construction.

I have solved this by creating a hierarchy of Config classes to
encapsulate the configuration data. This includes things like ensuring
the actor and plane strategy fields are set when calling
setBossStrategy() and having a flag tracking whether the position was
manually set. Other than streamlining object creation (one can see how
consistent the interface for things is across all places, and
especially the factories), this allows for only typing certain shared
parameters a single time - see: ScreenFactory.

I might want to refactor again to instead formally use a Singleton for
the controller (as I just realized that only one instance of it is ever
created) and stage (for similar reasons). This would reduce the argument
passing tedium even further.